🛠️refactor: edit useModal hook convention
