--- src/hooks/indext.ts
@@ -1,1 +1,0 @@
-export { default as useModal } from './useModal';
