+++ src/stories/components/AlertModal_Story/AlertModal.stories.tsx
@@ -1,5 +1,5 @@
 import { AlertModal } from '@/components/client';
-import useModal from '@/hooks/useModal';
+import { useModal } from '@/hooks';
 import type { Meta, StoryObj } from '@storybook/react';
 
 /**
